(fix) Add a first, unconditional conversion, to conversion flow
When optimizing rxjs flow we removed a fisrt conversion appearing
in original code. This is a first step wich is unnecessary if
converting by limiting file size, but necessary if converting by
limiting jpeg quality or image size. In addition, it has revealed
that quality and image size limits are never applied. This is a
trivial to implement functionality that has never been
implemented.