(feature) Add parameter to specify if resulting image must "cover" max size
In origin code, parameter maxImageSize in getScales meant
"reduce image size in order to cover completely the specified
maxImageSize". This was the needed behavior. But the "logic"
behavior is that maxImageSize means "max for the bigger image
dimension". So, we added an option to specify what does the user
meant each time with "maxImageSize"